pyproject.toml
[project]
name = "maia2-shogi"
version = "0.1.0"
description = "A set of tools for inference of Maia-2 Shogi"
license = "MIT"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cshogi",
    "japanize-matplotlib>=1.1.3",
    "matplotlib>=3.9.2",
    "onnx>=1.17.0",
    "onnxruntime-gpu>=1.20.1",
    "pandas>=2.2.3",
    "tqdm>=4.66.5",
]

[tool.uv.sources]
cshogi = { git = "https://github.com/nyoki-mtl/cshogi", rev = "numpy2-support" }

[tool.hatch.build.targets.wheel]
packages = [
    "src/maia2-shogi"
]

[dependency-groups]
dev = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.6.6",
    "pytest-asyncio>=0.26.0",
    "pytest>=8.3.4",
    "ruff>=0.9.9",
    "mkdocs-minify-plugin>=0.8.0",
    "mypy>=1.15.0",
]

[tool.ruff]
line-length = 120
exclude = ["_refs"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]

# Per-file rules
[tool.ruff.lint.per-file-ignores]
"scripts/**/*.py" = ["E402"]  # Allow imports after sys.path modifications in scripts

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 120

[tool.mypy]
# Pythonのバージョンを指定
python_version = "3.10"

# 型情報が不足しているライブラリに対する警告を抑制 (types-* スタブがない場合)
ignore_missing_imports = true

# _refsディレクトリを除外
exclude = ["_refs"]

# Any型を禁止するかどうか (段階的に導入する場合は最初はfalse推奨)
disallow_any_generics = false
disallow_untyped_defs = false # 型ヒントがない関数の定義を許可しない (最初はfalseで徐々にtrueへ)
